Fix the default tolerance being used by CHEBTECH/SIMPLIFY().
As pointed out by @gradwright in #294, the default tolerance being used was F.EPSLEVEL.*F.VSCALE, but the F.VSCALE is being included later.

The documentation states that if a TOL is passed, this takes the place of F.EPSLEVEL, so the F.VSSCALE in the conditional statement is the correct one to remove.

All tests pass. Closes #294.