Hack to deal with weird MATLAB access attribute bug???
chebfunpref.manageDefaultPrefs and cheboppref.manageDefaults are
declared private, as they probably should be.  This means that they
can't be accessed outside of chebfunpref and cheboppref unless we do so
through a function handle.  This is why chebpref.setDefaults takes a
manageSCDefaults function handle input.

Now, I can do

  >> chebfunpref.setDefaults('factory')
  >> chebfunpref.setDefaults('factory')
  >> chebfunpref.setDefaults('factory')
  ...

all day long, and there are no problems.  But if I make a file, say,
temp.m and put the following contents in it

  function temp()
      chebfunpref.setDefaults('factory')
      chebfunpref.setDefaults('factory')
  end

we get the following error:

  "Cannot access method 'manageDefaultPrefs' in class 'chebfunpref'."

This error will also emerge if I do

  >> chebfunpref.setDefaults('factory'), chebfunpref.setDefaults('factory')

It seems that if the MATLAB interpreter tries to compile two or more
lines of chebfunpref.setDefaults('factory') in the same run, something
weird happens, and we get a private method access error.

The solution implemented in this commit is to fool the MATLAB
interpreter by adding an extra level of indirection, wrapping our
function handle for chebfunpref.manageDefaultPrefs in an extra anonymous
function.  This is a hack, but then again, so is the deduplication
strategy employed in chebpref.setDefaults.  So I guess we get what's
coming to us in the end...