Support construction from vector epsilon inputs in chebtech.
The thing causing strictCheck() to break at the moment is related to vector
epslevels.  Specifically, in chebtech.compose(), we supply the epslevel of the
chebtech being composed to the constructor for determining tolerances, tail
lengths, etc.  The exact value being passed to the constructor when compose(f,
@op) is called is f.epslevel.  When f is array-valued, this is now a vector,
and strictCheck() is not set up to handle this.

Come to find out, the only reason classicCheck() wasn't breaking due to the
same problem was due to a combination of dumb luck and the fact that it was
written to treat the columns of an array-valued chebtech as if they depended on
one another.  We didn't originally conceive of vector-valued inputs for the
constructor epsilon when implementing vector epslevels previously, so this got
overlooked.  This commit modifies classicCheck() to decouple the columns and
construct each according to their own epsilon.  When constructing array-valued
chebtechs, you can now set pref.eps to be a vector and the "correct" thing will
happen.

Documentation discussing this new "feature" has been added to
@chebtech/happinessCheck.m but not @chebtech/techPref.m.  This is not a
capability that I think we want to advertise.

Tests pass with remarkably few changes.  We needed to loosen the tolerances on
two tests in chebfun/test_roots.m, but that's it.  For these, the error
increases were tiny:  the old tolerance was 7.0e-16, and the new error is
7.2e-16.

Similar changes in strictCheck() to follow.

This stuff is related to the issues raised in ticket #162.