Changes to fix #437.
The problem in #437 was that some breakpoints in a quasimatrix were close to each other, and when trying to merge them in order to create an array-valued CHEBFUN (which is the default strategy for computing QR of a quasimatrix when possible) tiny intervals were breaking RESTRICT().

This fix extends the TWEAKDOMAIN() method to support tweaking of quasimatrix breakpoints.

A new tets was added, and all existing tests pass.