Fix bug in chebfun.merge() for row chebfun inputs.
Inside chebfun.merge(f), we create the merged funs by calling the constructor
on @(x) feval(f, x).  The constructor expects that this function handle will
give column outputs for column inputs.  If f is a column chebfun, this holds,
but if f is a row chebfun, the output will be transposed.  Hence, we
un-transpose the input to chebfun.merge() at the top of the routine and
re-transpose it again at the bottom if needed.