Fixes for logical CHEBTECH methods.
These preiously only sampled at one interior point, however this can be easily fooled by sign changes at the level of rounding error, particularly when the CHEBTECH is representing a function on an unbounded domain through unbndfun.

To prevent this, we now same at the end points of the domain as well as the same arbitrary point, and take the mean.