Persitently store some matrices in @CHEBTECH/QR().
One of the biggest clients of @CHEBTECH/QR() is @CHEBFUN2/PLUS(), which often calls QR() consecutively with functions of the same length.

This fix stores the two weighted BARYMATS needed in QR for the previously used value of n to save computing it again if the next value of n is the same.