An attempt at more robust filtering of spurious roots in unbndfun.roots().
Previously, we set a threshold at around 10^15, and if the (magnitude of
the) location of a root was above this threshold, it was deemed likely
to be spurious, arising only due to the fact that the function being
represented decays rapidly at infinity.  Unfortunately, this criterion
isn't strict enough to handle this most of the time.  For instance, we
have following example from test 28 in tests/chebfun/test_norm():

  >> f = chebfun(@(x) (1 - exp(-x))./(x.^2), [1 Inf]);
  >> roots(f)
  ans =
     1.0e+07 *
     9.821248494834144
                   Inf

This commit tries to remedy this by checking to see if the function is
"essentially zero" on the interval between a root and +/- Inf and, if
so, decides that the root is spurious.  It does the check by restricting
f to the relevant interval and calling normest() to get a quick measure
of its size.  This is obviously expensive, but it seems to work, at
least for the preceding example:

  >> f = chebfun(@(x) (1 - exp(-x))./(x.^2), [1 Inf]);
  >> roots(f)
  ans =
     Inf

This does not yet work for array-valued unbndfuns.  I will implement
this if people decide that this is a good idea.