Move code to filter endpoint roots down into chebtech.
Technically, I suppose this belongs there, as all the problems which apply to
unbndfun apply equally to a chebtech formed from a function composed with a
suitable map.  This is also probably faster, as we can just do point evaluation
instead of having to call restrict().

We reinstate the old "far-field" filtering method in unbndfun, as it seems to
catch some cases that these changes miss.

Remarkably, all the tests pass, at least for me, but this still needs more
vetting before we can call it truly stable.