Make chebfunpref.subsref() smarter about looking for tech preferences.
Instead of just looking in prefList.techPrefs, which is not required to be a
complete list of tech preferences, we now get the _full_ list of tech
preferences by merging with the structure given by the currently selected
tech's techPref() function.  This allows you to do, e.g.,

  >> rf = chebfunpref().refinementFunction;

to get the default value of the refinement function if the current tech is
chebtech1 or chebtech2.

More subtly, you can also do

  >> p = chebfunpref();
  >> p.tech = @chebtech2;
  >> p.maxLength = 1000;

to set the maxLength preference.  You can then retrieve it via either the
abstract name

  >> maxlen = p.maxLength
  maxlen =
          1000

or via the more concrete name to which chebtech maps maxLength, i.e., maxPoints

  >> maxpts = p.maxPoints;
  maxpts =
          1000

At the moment, this does not work the other way, however:

  >> p.maxPoints = 2000;
  >> maxpts = p.maxPoints
  maxpts =
          2000
  >> maxlen = p.maxLength
  maxlen =
          1000

We may want to have the techPref functions return their name maps as a second
output to allow us to create an "inverse map" which would fix this apparent
discrepancy.