Reformat internal constructor prototypes.
In order to get all internal constructors to have the same prototype, we now
collect all non-op and non-pref parameters in a single struct called data.
Having them all have the same prototype makes it easier to call them abstractly
through methods like make(), as one no longer has to check the type of the
object on which make() is invoked in order to determine the constructor
parameters.

The work done here is sloppy in places and is meant to get this change working
as quickly as possible without regard for being exceptionally clean about it.
Polishing this up a bit is the next task.

This commit is very large, but the total number of substantive changes is
fairly small.  The vast majority (about 80%) of the changes consist of fixing
up the tests for the internal layers to respect the new conventions.

All tests in the system pass with these changes.