Clean up internal constructor logic.
Since we're reformatting the internal constructor prototypes, we may as
well take some time to clean up their insides as well.  This commit
attempts to simplify some of the logic in the constructors and just
generally clean things up a bit.

Probably the only significant change is that
pref.enableSinularityDetection now _must_ be enabled if you want to
activate singfun from within the lower layers.  You can't just set
pref.singPrefs.exponents or pref.singPrefs.singType (of which we are
ridding ourselves) and have singfun activated for you.