Construct chebfunpref objects with different base preferences.
Normally when you construct a chebfunpref, the preferences that get
overridden by any additional struct inputs are the defaults.  It turns
out, however, that it can be useful to allow the caller to supply the
base preferences to be overriden.  With the changes in this commit, the
user can now specify a chebfunpref argument to the chebfunpref
constructor to use as the base set.