Enable chebfun construction with both pref. objects and keyword arguments.
Since prefs.singPrefs.exponents and prefs.singPrefs.singTypes have been
eliminated, the only way to get the data they used to hold into the
constructor is via keyword arguments.  This makes life difficult if you
want to supply preferences to the constructor via a chebfunpref object
or MATLAB struct, since the constructor does not currently permit the
interspersion of preference objects and keyword arguments.

This commit changes things so that such interspersion is allowed.  The
user is allowed to supply at most one preference object argument.  In
the event that a preference supplied by keyword argument and a
preference supplied by the preference object conflict, the keyword
argument takes precedence.  Things like

  p = chebfunpref();
  p.tech = @chebtech1;
  f = chebfun(@(x) 1./x, [0 1], 'exps', [-1 0], pref);

now work just fine.

As a result of these changes, the input parsing logic for the chebfun
constructor has been overhauled quite a bit.  It's become
simpler/clearer in some places (e.g., how the domain is parsed) and less
simple/less clear in others (e.g., how singTypes get handled), so the
net change on the readability should be fairly small.  This function is
very complex and could use further overhaul, but that's not a top
priority right now.

All tests pass.  No chebfun tests required modification (aside, of
course, from the addition of a few tests to test these changes!), so all
syntaxes that are tested in the tests still work.