Pass data argument to internal compose() functions.
This allows us to get rid of the lone remaining reference to
pref.singPrefs.exponents in the test for singfun.compose().  All
internal compose functions now allow you to pass data arguments, so you
can, e.g., supply exponents to a composition this way instead of using
preferences.  Supplying them via preferences is no longer supported.

We have _not_ altered chebfun.compose() in a similar manner.  If we want
to allow this in chebfun as well, we need to think carefully about what
kind of syntax we would want to for it.    Moreover, we will need to
alter every method which calls compose (e.g., all the trig functions) to
accept arguments with this sort of syntax as well.  This is too much to
be bothered with for the time being, so we'll leave it alone.  Note that
v4 does not have this kind of feature, and hence there is no
backward-compatibility constraint.