Use filtered integrand in ubndnfun.sum().
This ought to help us with some of the accuracy problems we've been
having with unbndun.sum().  This effectively implements the filtering
scheme for doing these integrals which was used in Chebfun v4.  The idea
is that values of the function being integrated which are really small
are set _exactly_ to zero to prevent the magnifiction of rounding error
due to the blowup of the nonlinear map at the endpoints of [-1, 1].

The one oddity is that we use an exact-length construction for the
integrand.  This is a hack:  it only works for chebtech, and even then,
there's no real reason that it should.  We would like to do an adaptive
construction here, but that has proved difficult, as the fact that the
nonlinear map rescales very wide intervals near +/-Inf to very narrow
ones near +/-1 means that the filtered integrand exhibits sharp
transitions on Chebyshev grids of any reasonable size.

This commit is essentially a cleaner version of work started by KX
elsewhere, and many of the changes in this commit come from him.