Bypass singfun when constructing unbndfun integrands.
With the changes from the previous commit, unbndfun.sum() would
construct the integrand using g.onefun.make().  This means that if g is
a singfun, the singfun constructor gets called, which is not what we
want.  Since the integrand does not blow up, we don't need to do
singularity detection here; in fact, doing this causes one test to fail
due to the inaccuracies in the singularity detector.

This commit solves this by adding a 'techConstructor' option to the
get() methods of the interior layers (but not to chebfun).  A get(f,
'techConstructor') call is passed all the way down to the bottom, until
it hits a tech, which returns a handle to its own constructor.  This is
useful, non-encapsulation-breaking way to set pref.tech.  In our case,
since a tech is exactly what we want to construct, we don't bother with
preferences here and just invoke the constructor we get back directly.