Use unbndfun roots filtering only if "prune" is set.
Since this filtering is pretty specific to unbndfun (thus far, anyway),
we probably only want to use it when we know we need it.  So, following
a suggestion of Nick Hale, we activate it only when the "prune" option
for roots() is set and have unbndfun.roots() always set this option.