Add documentation to roots filtering function.
In particular, we note the reason why roots that are very near interval
endpoints are not pruned:  if we don't do this, then examples like
sin(500*x) will have their roots near the endpoints pruned, even though
they're legitimate.  On the other hand, in the case of unbndfun, these
roots probably _should_ be pruned, hence the retention of the
"far-field" filtering procedure in unbndfun.roots().

There are examples in the tests which trip up if either the skipping of
roots very near interval endpoints or the "far-field" filtering is
omitted, so although this is a subtle point, we are unlikely to miss it
in the future.