Make filtering of endpoints roots a separate roots preference.
This is better than piggybacking on top of "prune", which already has a
well defined behavior.  Moreover, it allows us to get rid of the
hack-ish blocks of code which don't filter roots that are deemed too
close to the endpoint.  If you find yourself losing roots near interval
endpoints, don't filter near them, plain and simple.