Add input to chebfun.get() for controlling output simplification.
As remarked in the help text, this is helpful for programmers, who generally
want the output to be as uniform as possible to make it easy to process without
requiring complicated logic.  The simplified versions are of more use to users
working in the command window, where unpacking cell arrays can be more
inconvenient.

This commit also greatly expands the documentation to spell out the exact rules
governing how things are simplified.