Modify ODE codes to work with new get().
The code which uses the outputs of these calls expects 2D cell array outputs in
situations in which get() will try to simplify them further.  So, we force
get() to return 2D cell arrays by setting the simplification level to 1.

With these changes, all existing tests pass on my machine.