Complete overhaul of *COEFFSPLOT() methods.
COEFFSPLOT() now exists. It simply organises the columns and funs of the inputs and calls the COEFFSPLOT() methods at the tech level where the actual plotting is done.

CHEBPOLYPLOT() is implemented as a trunk method which calls COEFFSPLOT() and throws a deprecation warning.

I thin the only thing not working is the LOGLOG option for FOURTECH()-based CHEBFUNS, because FOURTECH/COEFFSPLOT() uses three different function handles. This will need to be changed.