Fix get(f, 'exponents') when f has only one fun.
isvector() is the wrong thing to do here: if simpLevel > 2 and the
chebfun has only one fun, the exponents returned by
getArbitraryLocalProp will be a numeric vector, and you can't cell2mat()
that (not that we need to in this case anyway).

This also fixes a bug in the simplification logic for 'exponents':  the
conversion to numeric matrix was being performed even for simpLevel less
than 2.