Rename upper-layer preference:  "enableSingularityDetection" --> "blowup".
This was done with

  sed -i -e 's/enableSingularityDetection/splitting/' \
      $(grep -rl enableSingularityDetection *)

and a little manual tweaking to fix documentation and tests.

NB:  This rename does NOT mean that this preference behaves semantically like
the v4 "blowup" preference.  This _only_ controls whether or not singularity
detection is enabled, not what singularities are searched for.  The changes
required to update these semantics will be done separately.