Add support for legacy preferences in constructor calls.
We now support "splitdegree", "maxdegree", and "resampling" in
constructor calls.  Note that no support for these is available in
chebfunpref.setDefaults().  If you want to affect these quantities
persistently or as they are stored in preference objects, you need to
use the proper names "splitLength", "maxLength", and
"refinementFunction".

We've also added support for "splitLength" and "splitMaxLength", since
it seems weird to allows "splitdegree" but not its more suitably-named
cousins.  (Also, these second-tier preferenes are somewhat more likely
to be played with than the other ones on second-tiers.)