Fix bugs in chebfun.chebcoeffs() for second-kind coefficients.
There were two problems here.

First, the code hadn't been converted for the changes to the way we handle the
coefficients.  We never detected this because none of our tests exercised this
part of the code.  Some tests have now been added.

Second, there was an error which caused the last requested coefficients
(highest degree) to be computed incorrectly because we didn't calculate enough
first-kind coefficients to use the first-kind-to-second-kind recurrence to to
calculate these higher-order coefficients.  This doesn't matter if we ask for
all the second-kind coefficients of a fully-resolved function, since the
highest-degree coefficients are basically zero, but if we ask for only a few
coefficients or if the function is only piecewise smooth, it becomes obvious
that something is wrong.

We fix this by simply computing two extra first-kind coefficients when
second-kind ones are requested.  This should be OK for smooth functions, but
for piecwise-smooth ones which require the integrals, it will be more
expensive.  I've left a TODO in the file reminding us to revisit this if it
becomes a problem.