Revert chebfun.chebpoly() to use old format for outputting the coefficients.
This should fix the recent breaks to ATAP caused by reversing the order of the
coefficients.  (ATAP uses chebpoly(), not chebcoeffs() to do its work.)  This
function is deprecated anyway, so having it retain the old behavior doesn't
cost us much (though if it's deprecated, we should really consider axing it in
a future release, say, Chebfun v6.0).

Since retaining this behavior is important at least for the time being, I've
added a test (basically a copy of the test for chebcoeffs()) to prevent us from
messing things up.