Warn if no write permissions when executing Chebgui export tests.
David Watkins sent us an e-mail indicating that these tests were failing
on his machine, and while we don't have full confirmation, it looks like
this is due to the fact that he has Chebfun installed in a directory to
which his user account cannot write.  These tests now check if writing
is allowed and pass automatically with a warning if it isn't.