Deal with brekpoints in the periodic case.
In the following, what I mean by default discretization is the DISCRETIZATION field of CHEBOPPREF.
By default, it is CHEBCOLLOC2. But the user can change it typing
cheboppref.setDefaults('discretization', @anyDiscretizationTheUserWantsToUse).

Before, if using the 'periodic' tag, the discretization used was FOURCOLLOC, unless
the default discretization in CHEBOPPREF was ULTRAS (in this case ULTRAS was used).
Moreover, if the user was using a PREF object to solve the bvp, that is u = solvebvp(L, f, pref),
PREF.DISCRETIZATION was used.

Now, if using 'periodic', the discretization used is FOURCOLLOC, unless the default discretization is
ULTRAS (in this case ULTRAS is used), or if there is a breakpoint (in this case the default discretization is used, whatever it is).
It the user is using a PREF object, PREF.DISCRETIZATION is used.
If there is a breakpoint and the discretization used is FOURCOLLOC (either because FOURCOLLOC is the default discretization, or because the users passed a PREF object with FOURCOLLOC in PREF.DISCRETIZATION), we
throw an error.