Force equal endpoints for adjacent subintervals in bndfun.conv() output.
This was not a problem previously because each fun returned by
bndfun.conv() was added to the final chebfun being constructed inside
chebfun.conv() one at a time, so the minor discrepancies in the
endpoints got removed by chebfun.tweakDomain().  With the recent changes
to conv(), they're now all assembled into a chebfun before they get
added.  chebfun does not attempt to tweak the domains of the endpoints
if it is constructed directly from funs, so the resulting chebfun winds
up with some breakpoints that differ only by eps, and that causes
problems.  In particular, the geom/RoundingCorners.m and
stats/CentralLimitTheorem.m examples crash.

This commit fixes this by forcing adjacent endpoints of funs in the cell
array returned by bndfun.conv() to match exactly.