Fix issues with endpoints in deltafun.conv() prior to restriction.
The fix of the previous commit attempts to patch things up after the
call to restrict(), which seems to work at the moment but is probably
ultimately still prone to the same error we've been getting when trying
to run the stats/CentralLimitTheorem.m example we've been getting.  This
commit (1) sets the endpoints of each fun returned by bndfun.conv() to
come exactly from the set of breakpoints we get after coalescing ones
which are close to each other and (2) does so _before_ calling
restrict(), which should make everybody downstream happy.

Ugh... this issue has been very annoying.