"Properly" update epslevels after scalar-chebtech multiplcation.
This closes #1201.  The improper update to epslevel in chebtech.times() in
particular was causing rather strange results:

>> f = 1e16.*chebtech2(@(x) x)
f =
  2x1 chebtech2 array with properties:

      coeffs: [2x1 double]
      vscale: 1.000000000000000e+16
      hscale: 1
     ishappy: 1
    epslevel: 2.000000000000001

As a preemptive measure, I've gone ahead and changed several other places where
we were updating epslevel in this context based on our heuristic rules to use
updateEpslevel() instead as well.  There are also a few changes some things
having to do with how epslevel gets updated when a function with a zero vscale
is encountered to avoid NaNs.

I make no claim that these are the right things to do in these situations, only
that they are "reasonable".