Add a heuristic "bound" for updating the epslevel in cumsum().
Here's the "justification" for this:

Say f is the true function we want to approximate and p is the chebtech we end
up using to approximate it.  We sort of hope that we have something like

  ||f - p||/||f|| <= eps(p),

where eps(p) is the epslevel of p and ||.|| is the infinity norm.  If Ix is the
operator which integrates from -1 to x and I is the operator which integrates
from -1 to 1, we have

  |Ix(f) - Ix(p)| <= Ix(|f - p|)
                  <= I(|f - p|)
                  <= eps(p)*||f||*I(1)
                   = 2*eps(p)*||f||

If we replace ||f|| with the vscale of p and assume the vscale of Ix(f) is
reasonably well-approximated by the vscale of Ix(p), this suggests the "bound"

  eps(Ix(p)) <= 2*eps(p)*vscale(p) / vscale(Ix(p)).

No, this isn't rigorous, but then again, nothing epslevel-related is anyway.
At any rate, this at least should prevent epslevel explosion while still
allowing us to avoid the odd behavior pointed out in #1206.

To prevent problems similar to those from #1206 from arising in the future, I
want to change the default upper bounds in updateEpslevel() to Inf instead of
pref.eps, but I don't have a good understanding of how this will affect things.
Prior to this commit, making such a change resulted in problems with taking the
indefinite integral of, e.g., degree-2 polynomials, which have no tail, causing
the happinessCheck() in updateEpslevel() to return a new epslevel which is
complete nonsense.  After making this commit, it seems to be safe to do, but
I've decided instead to just leave TODOs there for now.

One tolerance on a tests in tests/bndfun/test_cumsum.m needed to be increased
slightly as a result of these changes.