"Handle" NaN chebtechs in chebtech.updateEpslevel().
tests/singfun/test_isnan.m does NaN*f in one place, and updateEpslevel() chokes
when it tries to compute the epslevel of the result because the NaN
coefficients cause classicCheck() to crash.  There are several possible ways to
handle this.  This commit does it by temporarily replacing the NaNs with zeros
to prevent classicCheck from crashing and then leaving the epslevels of NaN
columns unchanged at the end of the procedure.

I don't claim that this is the "correct" solution, since the whole "NaN
chebtech" concept has not been thought through anyway, but it should do for the
time being.