Fix bug in simplification of identically-zero chebtechs.
This addresses one of the points raised in #1283.  With the coefficients now
ordered from lowest degree to highest, if we want to leave just one coefficient
when our simplification yields an identically-zero result, we need to make that
we set the last nonzero coefficient row to the _first_ row, not the last.  This
modification should have been done a while ago but must have been missed when
the coefficient ordering was flipped.

A test has been added to check for this problem in the future.