Error out if given a keyword preference with no corresponding value.
Note that this only works for preferences which are lumped into
keywordPrefs and not those which are parsed separately.  Doing this for
the latter as well would probably double the length of the input parsing
routine with the way we're doing things currently.  To fix this, we
could rewrite the code for the preferences which are parsed separately
to use keywordPrefs.

That said, this is probably more complicated than it seems.  For one,
there are options like "trunc", which get handled in a different way
still, and for another, we sometimes need to do parsing on the value as
in "splitting" and "chebkind".  The way keywordPrefs works does not
allow for that at present, and it would not be advisable to make
substantial changes to the input parser so close to a release.

This is meant to address one of the points brought up in #1296.