Have (chebtech|trigtech).techPref() issue warnings for unknown preferences.
This commit is meant to address #1296.

There are two reasons this must be done inside chebtech and trigtech and
not inside chebfun.parseInputs().  First, and most importantly, techs
are the only objects which know enough to be able to validate their
preferences.  In this case, the validation is a simple check of struct
fieldnames and could easily be handled by chebfun by doing
pref.tech().techPref() at the appropriate part in the parsing routine
and comparing the fields to those in pref.techPrefs; however, one could
easily imagine a tech with a set of preferences which requires a more
complicated validation procedure, e.g., the preferences could have
nested structs, and it might be desired to do validation on the contents
of the substructs as well.

The disadvantage to doing this check at the bottom inside the individual
techs is that the overhead added could be greater than if we did a check
at the top in chebfun.parseInputs().  My current belief is that the
impact of this is likely to be minimal since this check is not done very
often (once per every tech constructed using its constructor when the
constructor is explicitly supplied a preference object), but we'll see.
Comparing the two strategies using a profiler run on chebtest('chebfun')
seems to indicate that running the check at the top has less overhead
but not by a whole lot, and there is no discernible difference between
the total time taken for their respective runs.

If the overhead turns out to be too great, an alternative might be to
create functions trigtech.validatePrefs() chebtech.validatePrefs() which
chebfun can call inside of chebfun.parseInputs() via something like
pref.tech().validatePrefs(inputPref).  The disadvantage to this is that
it won't work if funqui is used, but maybe that doesn't matter.

(Personally, I think that whether or not funqui is employed should be an
independent preference, i.e., it shouldn't piggyback on top of the tech
selection mechanism.  That would solve this problem, but this is a
discussion for another day.)