Fix some misspelled preference names.
Only two items to note here:

1.  The fact that we weren't aware of this until now means that the
impact these preference has on things seems to be unimportant.  Do we
really need to be setting them here?  If so, we should find some
examples demonstrating that and add them to the tests.

2.  chebfun really should not be messing with the refinementFunction
preference (or, equivalently, the "resampling" input to the
constructor).  The reason is that this is a preference that only makes
sense for chebtech/trigtech and may not apply to arbitrary techs (and
even if it does, it may not have the same meaning).  In most of the
cases where it's being used, it's setting it to use nested refinement,
which is already the default anyway.  In the one case where it sets it
to use non-nested refinement (@chebfun/inv.m), it has no bearing on the
outcome of the test, so perhaps we simply don't need to set it (as in
point #1, above).