Add ability to compute the transpose of the leg2cheb operator.
This is useful in computation of the inverse discrete Legendre transform, which we hope to add soon.

This commit also addresses some small bug fixes left over from the switch in order of coefficients in Chebfun, as well as some general tidying up.

Tests have been added for the new features.